+++ java/lang/StringIndexOutOfBoundsException/StringIndexOutOfBoundsException.cpp
@@ -40,12 +40,10 @@ StringIndexOutOfBoundsException::StringIndexOutOfBoundsException(Throwable *caus
 
 }
 
-StringIndexOutOfBoundsException::StringIndexOutOfBoundsException(String message, Throwable *cause)
-        : IndexOutOfBoundsException(message, cause) {
+StringIndexOutOfBoundsException::StringIndexOutOfBoundsException(String message, Throwable *cause) : IndexOutOfBoundsException(message, cause) {
 
 }
 
-StringIndexOutOfBoundsException::StringIndexOutOfBoundsException(long int index) : IndexOutOfBoundsException(
-        String("String index out of range: ") + String::valueOf(index)) {
+StringIndexOutOfBoundsException::StringIndexOutOfBoundsException(int index) : IndexOutOfBoundsException(String("String index out of range: ") + String::valueOf(index)) {
 
 }
